tree-construction/main-element.dat

#data
<!doctype html><p>foo<main>bar<p>baz
#errors
(1,36): expected-closing-tag-but-got-eof
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <p>
|       "foo"
|     <main>
|       "bar"
|       <p>
|         "baz"

#data
<!doctype html><main><p>foo</main>bar
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <main>
|       <p>
|         "foo"
|     "bar"
